dice, choose: extract `.choose` command into own module
Could have moved it to `rand.py`, but it doesn't *really* fit in any
better there. There's very little cost to having a standalone file for
this command aside from possibly breaking exclude settings in existing
Sopel installations (and we're already adding other new modules in 7.0).

Made sure this new file credits all the right people. Namely, I added
Exirel to the header, because he did a significant chunk of work on this
function last year.
